+++ juwonlee/board-project/src/components/week9/task2/Task2.js
@@ -1,12 +1,6 @@
-import { useState, useEffect } from 'react';
+import { useCounter } from './useCounter.js';
 
 export default function Task2() {
-  const [count, setCount] = useState(0);
-  useEffect(() => {
-    const id = setInterval(() => {
-      setCount(c => c + 1);
-    }, 1000);
-    return () => clearInterval(id);
-  }, []);
+  const count = useCounter();
   return <h1>Seconds passed: {count}</h1>;
 }
